+++ lib/repository.rb
@@ -4,9 +4,24 @@ class Repository
     rows = CSV.open file, headers: true, header_converters: :symbol
     rows.each do |row|
       new_obj = class_var.new(row.to_h)
-      # require 'pry'; binding.pry
       @repo << new_obj
     end
+  end
+
+  def new_id(attributes)
+    unless @repo.empty?
+      attributes[:id] = all.max do |item|
+        item.id
+      end.id + 1
+    end
+  end
+
+  def find_by_id(id)
+    @repo.find { |item| item.id == id }
+  end
+
+  def delete(id)
+    @repo.delete_if { |item| item.id == id }
   end
 
   def inspect
